Extract calendar component
